"build: create utils.js "
